demo/graph_tutorial/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.2
idna==2.10
msal==1.10.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
openpyxl==2.5.1